:recycle: GI-5 Refactor stocks balance
