+++ src/stocks_balance.py
@@ -10,15 +10,7 @@ from message_queue_events import StocksBalanceEvent
 from services import message_queue
 from services.converters import UnitsConverter
 from services.external_dodo_api import DatabaseAPI, DodoAPI, AuthAPI
-
-
-def group_stocks_balance_by_unit_id(
-        stocks_balance: Iterable[models.UnitStocksBalance],
-) -> DefaultDict[int, list[models.UnitStocksBalance]]:
-    unit_id_to_stocks_balance = collections.defaultdict(list)
-    for stock_balance in stocks_balance:
-        unit_id_to_stocks_balance[stock_balance.unit_id].append(stock_balance)
-    return unit_id_to_stocks_balance
+from shortcuts.stocks_balance import get_stocks_balance, group_stocks_balance_by_unit_id
 
 
 def main():
@@ -29,18 +21,11 @@ def main():
         units = DatabaseAPI(database_client).get_units()
     units = UnitsConverter(units)
 
-    stocks_balance: list[models.UnitStocksBalance] = []
     with httpx.Client(base_url=config.api.auth_api_base_url) as auth_client:
         with httpx.Client(base_url=config.api.dodo_api_base_url, timeout=60) as dodo_api_client:
             auth_api = AuthAPI(auth_client)
             dodo_api = DodoAPI(dodo_api_client)
-            for account_name, grouped_units in units.grouped_by_account_name.items():
-                account_cookies = auth_api.get_account_cookies(account_name)
-                stocks_balance += dodo_api.get_stocks_balance(
-                    unit_ids=grouped_units.ids,
-                    cookies=account_cookies.cookies,
-                    days_left_threshold=1,
-                ).units
+            stocks_balance = get_stocks_balance(dodo_api=dodo_api, auth_api=auth_api, units=units)
 
     events = [
         StocksBalanceEvent(
